Fix static checks
